Refactor pageLinks into PageLink component
* Moved pageLinks markup into PageLink component
