+++ pages/index.jsx
@@ -6,53 +6,37 @@ import DocumentTitle from 'react-document-title'
 import { prefixLink } from 'gatsby-helpers'
 import access from 'safe-access'
 import { config } from 'config'
-import SitePost from '../components/SitePost'
 import SiteSidebar from '../components/SiteSidebar'
+import PageLink from '../components/PageLink'
 
 class SiteIndex extends React.Component {
-    render() {
-        const pageLinks = []
-        // Sort pages.
-        const sortedPages = sortBy(this.props.route.pages, (page) => access(page, 'data.date')
-        ).reverse()
-        sortedPages.forEach((page) => {
-            if (access(page, 'file.ext') === 'md' && access(page, 'data.layout') === 'post') {
-                const title = access(page, 'data.title') || page.path
-                const description = access(page, 'data.description')
-                const datePublished = access(page, 'data.date')
-                const category = access(page, 'data.category')
+  render() {
+    const sortedPages = sortBy(this.props.route.pages, (page) => 
+      access(page, 'data.date')).reverse()
 
-                pageLinks.push(
-                    <div className='blog-post' key={title}>
-                      <time dateTime={ moment(datePublished).format('MMMM D, YYYY') }>
-                        { moment(datePublished).format('MMMM YYYY') }
-                      </time>
-                      <span style={ {    padding: '5px'} }></span>
-                      <span className='blog-category'>{ category }</span>
-                      <h2><Link style={ {    borderBottom: 'none',} } to={ prefixLink(page.path) } > { title } </Link></h2>
-                      <p dangerouslySetInnerHTML={ {    __html: description} } />
-                      <Link className='readmore' to={ prefixLink(page.path) }> Read
-                      </Link>
-                    </div>
-                )
-            }
-        })
-
-        return (
-            <DocumentTitle title={ config.siteTitle }>
-              <div>
-                <SiteSidebar {...this.props}/>
-                <div className='content'>
-                  <div className='main'>
-                    <div className='main-inner'>
-                      { pageLinks }
-                    </div>
-                  </div>
-                </div>
-              </div>
-            </DocumentTitle>
-        )
+    function isMarkdownPost(page) {
+      return(
+        access(page, 'file.ext') === 'md' 
+        && access(page, 'data.layout') === 'post'
+      )
     }
+
+    return (
+      <DocumentTitle title={ config.siteTitle }>
+        <div>
+          <SiteSidebar {...this.props}/>
+          <div className='content'>
+            <div className='main'>
+              <div className='main-inner'>
+                { sortedPages.filter(isMarkdownPost).map((page) =>
+                  <PageLink key={page.data.path} page={page} />) }
+              </div>
+            </div>
+          </div>
+        </div>
+      </DocumentTitle>
+    )
+  }
 }
 
 SiteIndex.propTypes = {
